Included fix from 3.0 to remove cast warning on some systems.
git-svn-id: file:///svn/nsd/branches/NSD_2_2@2542 a26ef69c-88ff-0310-839f-98b793d9c207

+++ util.c
@@ -92,7 +92,7 @@ log_file(int priority, const char *message)
 	
 	/* Bug #104, add time_t timestamp */
 	fprintf(current_log_file, "[%d] %s[%d]: %s: %s",
-		time(NULL), global_ident, (int) getpid(), priority_text, message);
+		(int)time(NULL), global_ident, (int) getpid(), priority_text, message);
 	length = strlen(message);
 	if (length == 0 || message[length - 1] != '\n') {
 		fprintf(current_log_file, "\n");
